extract checkbox components into separate modules
